move get_spectra_by_element to calcfunction
